backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
cryptography==42.0.8

# Background Tasks
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0

# SSH & Remote Management
paramiko==3.4.0
ansible-runner==2.3.4

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pytz==2023.3
Pillow==10.1.0
croniter==1.4.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0